+++ src/objects/index.js
@@ -4,9 +4,8 @@
 export { createBall, createBeachBall } from "./Balls";
 export { createBox, floydWords } from "./Boxes";
 export { createBillboard, createBillboardRotated } from "./Billboards";
-export { createWallX, createWallZ, wallOfBricks, createBrick } from "./Walls";
+export { createWall, wallOfBricks } from "./Walls";
 export { createGridPlane } from "./Planes";
 export { createTriangle } from "./Shapes";
 export { loadFloydText, loadEngineerText } from "./Text";
 export { addRigidPhysics } from "./PhysicsHelpers";
-