+++ src/utils/index.ts
@@ -1,7 +1,7 @@
 /**
  * 工具函数统一导出
  */
-export { loadTexture, type TextureLoadOptions } from "./textureLoader";
+export { loadTexture } from "./textureLoader";
 export {
 	getRandomArbitrary,
 	getRandomInt,